% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_shp.R
\name{zip_shp}
\alias{zip_shp}
\title{zip_shp()}
\usage{
zip_shp(file_name, zip_pattern = ".zip$", shp_pattern = ".shp$")
}
\arguments{
\item{file_name}{nombre del archivo, obligatorio}

\item{zip_pattern}{formato del archivo comprimido}

\item{shp_pattern}{formato del shp}
}
\value{
.shp
}
\description{
Función que permite leer un shp contenido en un zip
}
