% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_shp.R
\name{crear_shp}
\alias{crear_shp}
\title{crear_shp()}
\usage{
crear_shp(
  shp_data,
  criterio_seleccion,
  agrupar_por,
  columnas_adicionales = NULL
)
}
\arguments{
\item{shp_data}{datos del shp}

\item{criterio_seleccion}{vector o lista con nombres, donde el nombre de cada elemento es el nombre de la columna y el valor es el valor a seleccionar}

\item{agrupar_por}{columnas por las que se agrupará el shp}

\item{columnas_adicionales}{columnas adicionales a incluir en el shp}
}
\value{
.shp
}
\description{
Función que permite generar un nuevo shp según los parámetros especificados
}
